third_party/blink/web_tests/fast/dom/shadow/shadow-root-direction-expected.txt

CONSOLE WARNING: line 8: Element.createShadowRoot is deprecated and will be removed in M80, around February 2020. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
Element direction details are accessible to shadow root

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(shadowRoot.getElementById("target2")).direction is "rtl"
PASS getComputedStyle(root).direction is "rtl"
PASS successfullyParsed is true

TEST COMPLETE
